refactor(testgen): extract selector healing into internal/testgen/heal
internal/testgen held three independent feature engines behind one package
boundary: test generation, failure classification, and selector healing. The
reference graph shows zero edges between the heal cluster and the rest — no
heal function is called from generate.go/classify.go/helpers.go and vice versa
— so heal moves out as a self-contained subpackage that imports nothing from
Kaboom at all.

The heal-only types and constants (TestHealRequest, HealResult, HealedSelector,
HealSummary, BatchHealResult, FileHealResult, the three selector/file error
codes, and the four batch limits) move with it; testgen keeps only the
generation and classification contracts.

Note on evidence: the heal_* filename prefixes were all created by a single
splitting commit (46b16919), so "the filenames agree with the history" is one
fact counted twice. The grouping here rests on the call graph, which is an
independent signal and points the same way.

internal/testgen: 11 -> 5 source files (drops out of the folder-size baseline)
internal/testgen/heal: 7 source files

No behavior change: exported identifiers, JSON field names, and error strings
are byte-identical; only the import path moved. heal_test.go moves with the
code it exercises (heal package coverage 86.0%, no function at 0.0%).
