+++ .folder-size-baseline.json
@@ -12,7 +12,6 @@
     "internal/mcp": 12,
     "internal/recording": 16,
     "internal/schema": 17,
-    "internal/testgen": 11,
     "internal/tools/analyze": 13,
     "internal/tools/configure": 15,
     "internal/tools/observe": 32,
