+++ docs/features/feature/test-generation/index.md
@@ -13,6 +13,16 @@ code_paths:
   - cmd/browser-agent/internal/testgenhandler/heal.go
   - cmd/browser-agent/internal/testgenhandler/generate.go
   - cmd/browser-agent/tools_generate.go
+  - internal/testgen/generate.go
+  - internal/testgen/helpers.go
+  - internal/testgen/classify.go
+  - internal/testgen/types.go
+  - internal/testgen/heal/batch.go
+  - internal/testgen/heal/paths.go
+  - internal/testgen/heal/repair.go
+  - internal/testgen/heal/selectors.go
+  - internal/testgen/heal/summary.go
+  - internal/testgen/heal/types.go
   - internal/schema/generate.go
 test_paths:
   - cmd/browser-agent/internal/testgenhandler/context_test.go
@@ -21,6 +31,8 @@ test_paths:
   - cmd/browser-agent/internal/testgenhandler/classify_test.go
   - internal/testgen/generate_test.go
   - internal/testgen/helpers_test.go
+  - internal/testgen/classify_test.go
+  - internal/testgen/heal/heal_test.go
   - internal/schema/invariants_test.go
 last_verified_version: 0.7.12
 last_verified_date: 2026-03-05
@@ -54,6 +66,8 @@ last_verified_date: 2026-03-05
 - Alias/contracts: `cmd/browser-agent/internal/testgenhandler/aliases.go`
 - Provider delegation: `cmd/browser-agent/internal/testgenhandler/provider_adapter.go`
 - Heal and classify handlers: `cmd/browser-agent/internal/testgenhandler/heal.go`, `cmd/browser-agent/internal/testgenhandler/classify.go`
+- Test generation and failure classification engine: `internal/testgen/`
+- Selector healing engine (self-contained subpackage): `internal/testgen/heal/`
 - Generate tool schema contract: `internal/schema/generate.go`
 - Core behavior tests: `cmd/browser-agent/internal/testgenhandler/context_test.go`, `cmd/browser-agent/internal/testgenhandler/generate_test.go`, `cmd/browser-agent/internal/testgenhandler/heal_test.go`, `cmd/browser-agent/internal/testgenhandler/classify_test.go`
 - Schema invariants: `internal/schema/invariants_test.go`
