+++ internal/testgen/heal/paths.go
@@ -1,6 +1,6 @@
 // Purpose: Validates test file and directory paths against project boundaries.
 // Why: Isolates path security validation from selector healing logic.
-package testgen
+package heal
 
 import (
 	"fmt"
