+++ internal/testgen/heal/repair.go
@@ -1,6 +1,6 @@
 // Purpose: Repairs individual broken CSS selectors using heuristic rewrite strategies.
 // Why: Separates single-selector repair logic from batch processing and path validation.
-package testgen
+package heal
 
 import (
 	"fmt"
