+++ internal/testgen/heal/selectors.go
@@ -1,6 +1,6 @@
 // Purpose: Detects dangerous patterns in selectors and finds selectors in test file content.
 // Why: Separates selector scanning and safety checks from repair and batch logic.
-package testgen
+package heal
 
 import (
 	"fmt"
