+++ internal/testgen/heal/summary.go
@@ -1,6 +1,6 @@
 // Purpose: Formats human-readable summaries of test healing results for MCP responses.
 // Why: Separates presentation formatting from healing computation and selector analysis.
-package testgen
+package heal
 
 import "fmt"
